Update remote-id-attribute default and description.
HTTP_OIDC_ISS is upstream's default and better suited to be used as the
default in the charm as well.

iss: Issuer Identifier for the Issuer of the response. The iss value is
a case sensitive URL using the https scheme that contains scheme, host,
and optionally, port number and path components and no query or
fragment components.